requirements.txt
streamlit==0.83.0
tensorflow==2.5.0
pandas==1.2.5
tensorflow_hub==0.12.0
numpy==1.21.0
Pillow==8.3.1
